music_translation_backend/requirements.txt
Flask==3.0.0
uvicorn==0.27.1
librosa==0.10.2.post1
# pretty-midi==0.2.10
# numpy>=1.22.0
# scipy>=1.10.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==7.4.3
flask-cors==4.0.0
basic-pitch